requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
sentence-transformers==5.0.0
faiss-cpu==1.11.0
numpy==2.3.0
requests==2.32.4
spacy==3.8.7 